--- src/Microsoft.Data.SqlClient/netcore/src/Common/tests/System/Collections/DictionaryExtensions.cs
@@ -1,20 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace System.Collections.Generic
-{
-    internal static class DictionaryExtensions
-    {
-        public static bool TryAdd<TKey, TValue>(this IDictionary<TKey, TValue> dict, TKey key, TValue value)
-        {
-            if (!dict.ContainsKey(key))
-            {
-                dict.Add(key, value);
-                return true;
-            }
-
-            return false;
-        }
-    }
-}
