+++ src/Microsoft.Data.SqlClient/tests/ManualTests/SQL/Common/InternalConnectionWrapper.cs
@@ -10,6 +10,22 @@ using Microsoft.Data.SqlClient.ManualTesting.Tests.SystemDataInternals;
 
 namespace Microsoft.Data.SqlClient.ManualTesting.Tests
 {
+#if NETFRAMEWORK
+    internal static class DictionaryExtensions
+    {
+        public static bool TryAdd<TKey, TValue>(this IDictionary<TKey, TValue> dict, TKey key, TValue value)
+        {
+            if (!dict.ContainsKey(key))
+            {
+                dict.Add(key, value);
+                return true;
+            }
+
+            return false;
+        }
+    }
+#endif
+
     public class InternalConnectionWrapper
     {
         private static Dictionary<string, string> s_killByTSqlConnectionStrings = new Dictionary<string, string>();
